third_party/blink/web_tests/virtual/xslt-disabled/fast/xsl/nbsp-in-stylesheet-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been removed in this browser. See https://chromestatus.com/feature/4709671889534976.
Test for bug 19215: XSLT transformToDocument fails when stylesheet contains &#160;

Exception: NotSupportedError: Failed to construct 'XSLTProcessor': XSLT is disabled
